fix: resolve all Xcode 26 / Swift 6 build errors (BUILD SUCCEEDED)
- SettingsStore: replace @AppStorage with UserDefaults+didSet (@Observable incompatibility)
- SwiftBridgeCore: add bridging header instead of SWIFT_INCLUDE_PATHS (RustStr not in scope)
- SwiftBridgeCore: make ToRustStr.toRustStr rethrows to accept throwing closures
- bridge-ffi: add `try` to bridge_open_database's toRustStr call
- bridge-ffi: add Swift.Error + @unchecked Sendable to BridgeFfiError hierarchy
- ScanPipeline: use enumerator.allObjects instead of for-in (makeIterator unavailable in async)
- ThumbnailGridView: .ignore → .ignored (KeyPress.Result API change in macOS 26)
- BridgeCore: rustVec[UInt(i)] → rustVec[i] (subscript takes Int not UInt)
- AboutView: .accent → .tint (ShapeStyle.accent removed in macOS 26)

+++ xcode/BridgeLite/BridgeLite/Bridging/BridgeCore.swift
@@ -199,7 +199,7 @@ extension Data {
     init(rustVec: RustVec<UInt8>) {
         let len = Int(rustVec.len())
         var bytes = [UInt8](repeating: 0, count: len)
-        for i in 0..<len { bytes[i] = rustVec[UInt(i)] }
+        for i in 0..<len { bytes[i] = rustVec[i] }
         self.init(bytes)
     }
 }
